use ceill instead of ceil

+++ src/sundials/stl/sunstl_vector.h
@@ -97,7 +97,7 @@ static inline SUNErrCode MAKE_NAME(SUNStlVectorTtype, Grow)(SUNStlVectorTtype se
        We explicitly cast capacity to a long double to silence any implicit
        conversion compiler warning. */
     int64_t new_capacity =
-      (int64_t)(ceil(((long double)self->capacity) * GROWTH_FACTOR));
+      (int64_t)(ceill(((long double)self->capacity) * GROWTH_FACTOR));
     return MAKE_NAME(SUNStlVectorTtype, Reserve)(self, new_capacity);
   }
   return SUN_SUCCESS;
